Test package refactored
